File appending and better error checking

+++ lib/WebAppCommon/Util/FileAccess.pm
@@ -28,26 +28,42 @@ sub _build_scp {
 sub get_file_content {
     my ( $self, $path ) = @_;
     my $dest = tmpnam();
-    $self->scp->get($path, $dest);
-    my $content = read_file($dest);
-    unlink $dest;
-    return $content;
+    my $success = $self->scp->get($path, $dest);
+    $self->log->debug($self->scp->{errstr}) if !$success;
+    my $want_array = wantarray;
+    if ( -e $dest ) {
+        my $content = read_file($dest, array_ref => $want_array);
+        unlink $dest;
+        return $want_array ? @{$content} : $content;
+    }
+    else {
+        return undef;
+    }
 }
 
 sub post_file_content {
     my ( $self, $path, $content ) = @_;
+    $self->log->debug("posting file content to $path");
     my $src = tmpnam();
     write_file($src, $content);
-    $self->scp->put($src, $path);
-    $self->log->debug("posting file content to $path");
+    my $success = $self->scp->put($src, $path);
+    $self->log->debug($self->scp->{errstr}) if !$success;
     unlink $src;
-    return;
+    return !$success;
+}
+
+sub append_file_content {
+    my ( $self, $path, $content ) = @_;
+    my $existing = $self->get_file_content($path) // q//;
+    return $self->post_file_content($path, $existing . $content);
 }
 
 sub make_dir {
     my ( $self, $path ) = @_;
     $self->log->debug("posting directory to $path");
-    $self->scp->mkdir($path);
+    my $success = $self->scp->mkdir($path);
+    $self->log->debug($self->scp->{errstr}) if !$success;
+    return !$success;
 }
 
 1;
